fix: resolve type mismatch in DATEDIF month calculation

+++ app/Formula/Builtins.hs
@@ -287,7 +287,7 @@ apply3 DateDiffOp d1 d2 unit = case (dateVal d1, dateVal d2, text unit) of
       'm' ->
         let (ya, ma, _) = toGregorian a
             (yb, mb, _) = toGregorian b
-         in VNum (fromIntegral ((yb - ya) * 12 + (mb - ma)))
+         in VNum (fromIntegral (yb - ya) * 12 + fromIntegral (mb - ma))
       'd' -> VNum (fromIntegral (diffDays b a))
       _ -> VErr "DATEDIF unit must be \"Y\", \"M\", or \"D\""
     [] -> VErr "DATEDIF unit must be \"Y\", \"M\", or \"D\""
